+++ tests/server/acl_tests.js
@@ -21,7 +21,7 @@ limitations under the License.
     var expect = chai.expect;
 
     Munit.run({
-        name: "Munit Helpers ACL Tests",
+        name: "munit-helpers - Server - ACL",
 
         tearDown: MunitHelpers.restoreAll,